requirements.txt
flask 
flask-smorest
flask-cors
biopython